requirements.txt
langchain
langchain-community
langchain-chroma
langchain-huggingface
langchain-openai
gradio
sentence-transformers
rank-bm25
transformers
torch
torchvision
torchaudio
chromadb
numpy
pandas
jupyter
nbformat
pymupdf
python-dotenv
openai
dateparser
pillow
# RAG 평가 시각화 의존성
matplotlib
seaborn
plotly
kaleido
scikit-learn